requirements.txt
environs==9.2.0
Django==3.1.4
dj_database_url==0.5.0
djangorestframework==3.12.2
